adding tests

+++ library_management/views.py
@@ -1,14 +1,7 @@
-import json
 from datetime import datetime
 
-from django.shortcuts import render, redirect
-from django.http import HttpResponse
-from django.http import JsonResponse
-from django.core import serializers as core_serializers
-
-from . import serializers
 from .models import Book, Student, BookIssue
-from .serializers import BookSerializer, BookIssueSerializer, StudentSerializer, FilterBookSerializer
+from .serializers import BookSerializer, BookIssueSerializer, StudentSerializer
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
@@ -21,27 +14,7 @@ def view_all_books(request):
     if request.method == 'GET':
         books = Book.objects.all()
         serializer = BookSerializer(books, many=True)
-        return JsonResponse(serializer.data, safe=False)
-
-
-@api_view(['GET', 'PUT', 'DELETE'])
-def get_book_details_by_id(request, id):
-    try:
-        book = Book.objects.get(pk=id)
-    except Book.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-    if request.method == 'GET':
-        serializer = BookSerializer(book)
-        return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = BookSerializer(book, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    elif request.method == 'DELETE':
-        book.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 @api_view(['POST'])
@@ -68,15 +41,21 @@ def add_student(request):
 def view_all_students(request):
     students = Student.objects.all()
     serializer = StudentSerializer(students, many=True)
-    return JsonResponse(serializer.data, safe=False)
+    return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 @api_view(['POST'])
 def check_out(request):
     book_id = request.data['book_id']
     student_id = request.data['student_id']
-    book = Book.objects.get(id=book_id)
-    student = Student.objects.get(id=student_id)
+    try:
+        book = Book.objects.get(pk=book_id)
+    except Book.DoesNotExist:
+        return Response({"Message": "Book not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+    try:
+        student = Student.objects.get(id=student_id)
+    except Student.DoesNotExist:
+        return Response({"Message": "Student not found in DB"}, status=status.HTTP_404_NOT_FOUND)
     if book.quantity_available >= 1:
         book.quantity_available -= 1
         book_issued = BookIssue()
@@ -86,51 +65,61 @@ def check_out(request):
         book_issued.quantity_issued += 1
         book.save()
         book_issued.save()
-        return Response({"message": "Book issued successfully"}, status=status.HTTP_201_CREATED)
-    return Response({"error": "quantity not available"})
+        return Response({"Message": "Book issued successfully"}, status=status.HTTP_201_CREATED)
+    return Response({"Error": "Quantity not available"}, status=status.HTTP_404_NOT_FOUND)
 
 
 @api_view(['PUT'])
 def check_in(request):
     book_id = request.data['book_id']
     student_id = request.data['student_id']
-    book = Book.objects.get(id=book_id)
-    if book is None:
-        return Response({"Error": "Invalid book id"})
-    student = Student.objects.get(id=student_id)
-    if student is None:
-        return Response({"Error": "Invalid student id"})
-    book_issued = BookIssue.objects.filter(book_id=book_id, student_id=student_id, status="Borrowed").first()
-    if book_issued is None:
-        return Response({"Error": "No books issued currently"})
+    try:
+        book = Book.objects.get(pk=book_id)
+    except Book.DoesNotExist:
+        return Response({"Message": "Book not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+    try:
+        student = Student.objects.get(id=student_id)
+    except Student.DoesNotExist:
+        return Response({"Message": "Student not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+
+    try:
+        book_issued = BookIssue.objects.filter(book_id=book_id, student_id=student_id, status="Borrowed").first()
+    except BookIssue.DoesNotExist:
+        return Response({"Error": "No books issued currently"}, status=status.HTTP_404_NOT_FOUND)
     book_issue_object = BookIssue.objects.get(id=book_issued.id)
     book_issue_object.status = "Returned"
+    book_issue_object.return_date = datetime.today().date()
     book_issue_object.quantity_issued -= 1
     book.quantity_available += 1
     book.save()
     book_issue_object.save()
-    return Response({"Message": "Book returned successfully"})
+    return Response({"Message": "Book returned successfully"}, status=status.HTTP_200_OK)
 
 
 @api_view(['PUT'])
 def update_checkin(request):
     return_book_id = request.data['return_book_id']
     new_book_id = request.data['new_book_id']
+    try:
+        return_book = Book.objects.get(pk=return_book_id)
+    except Book.DoesNotExist:
+        return Response({"Message": "Return Book not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+    try:
+        new_book = Book.objects.get(pk=new_book_id)
+    except Book.DoesNotExist:
+        return Response({"Message": "New Book not found in DB"}, status=status.HTTP_404_NOT_FOUND)
     student_id = request.data['student_id']
-    return_book = Book.objects.get(id=return_book_id)
-    new_book = Book.objects.get(id=new_book_id)
-    if return_book is None:
-        return Response({"Error": "Invalid return book id"})
-    student = Student.objects.get(id=student_id)
-    if student is None:
-        return Response({"Error": "Invalid student id"})
-    if new_book is None:
-        return Response({"Error": "Invalid issue book id"})
-    book_issued = BookIssue.objects.filter(book_id=return_book_id, student_id=student_id, status="Borrowed").first()
-    if book_issued is None:
-        return Response({"Error": "No books issued currently"})
+    try:
+        student = Student.objects.get(id=student_id)
+    except Student.DoesNotExist:
+        return Response({"Message": "Student not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+    try:
+        book_issued = BookIssue.objects.filter(book_id=return_book_id, student_id=student_id, status="Borrowed").first()
+    except BookIssue.DoesNotExist:
+        return Response({"Error": "No books issued currently"}, status=status.HTTP_404_NOT_FOUND)
     book_issue_object = BookIssue.objects.get(id=book_issued.id)
     book_issue_object.status = "Returned"
+    book_issue_object.return_date = datetime.today().date()
     book_issue_object.quantity_issued -= 1
     return_book.quantity_available += 1
     return_book.save()
@@ -144,52 +133,92 @@ def update_checkin(request):
         book_issued.quantity_issued += 1
         new_book.save()
         book_issued.save()
-    return Response({"message": "Book returned and issued successfully"}, status=status.HTTP_200_OK)
+    return Response({"Message": "Book returned and issued successfully"}, status=status.HTTP_200_OK)
 
 
-@api_view(['DELETE'])
-def delete_book(request, book_id):
+@api_view(['GET', 'PUT', 'DELETE'])
+def get_book_details_by_id(request, id):
     try:
-        book = Book.objects.get(pk=book_id)
+        book = Book.objects.get(pk=id)
     except Book.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-    book.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response({"Message": "Book not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+    if request.method == 'GET':
+        serializer = BookSerializer(book)
+        return Response(serializer.data)
+    elif request.method == 'PUT':
+        serializer = BookSerializer(book, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'DELETE':
+        book.delete()
+        return Response({"Message": "Book with given id deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
+
+
+@api_view(['GET', 'PUT', 'DELETE'])
+def get_student_details_by_id(request, id):
+    try:
+        student = Student.objects.get(pk=id)
+    except Student.DoesNotExist:
+        return Response({"Message": "Student not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+    if request.method == 'GET':
+        serializer = StudentSerializer(student)
+        return Response(serializer.data)
+    elif request.method == 'PUT':
+        serializer = StudentSerializer(student, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'DELETE':
+        student.delete()
+        return Response({"Message": "Student with given id deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['DELETE'])
-def delete_student(request, student_id):
+def delete_book(request, id):
     try:
-        student = Student.objects.get(pk=student_id)
+        book = Book.objects.get(pk=id)
+    except Book.DoesNotExist:
+        return Response({"Message": "Book not found in DB"}, status=status.HTTP_404_NOT_FOUND)
+    book.delete()
+    return Response({"Message": "Book deleted successfully"}, status=status.HTTP_200_OK)
+
+
+@api_view(['DELETE'])
+def delete_student(request, id):
+    try:
+        student = Student.objects.get(pk=id)
     except Student.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+        return Response({"Message": "Student not found in DB"}, status=status.HTTP_404_NOT_FOUND)
     student.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    return Response({"Message": "Student deleted successfully"}, status=status.HTTP_200_OK)
 
 
 @api_view(['PUT'])
-def update_book_details(request, book_id):
+def update_book_details(request, id):
     try:
-        book = Book.objects.get(pk=book_id)
+        book = Book.objects.get(pk=id)
     except Book.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+        return Response({"Message": "Book not found in DB"}, status=status.HTTP_404_NOT_FOUND)
     serializer = BookSerializer(book, data=request.data)
     if serializer.is_valid():
         serializer.save()
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['PUT'])
-def update_student_details(request, student_id):
+def update_student_details(request, id):
     try:
-        student = Student.objects.get(pk=student_id)
+        student = Student.objects.get(pk=id)
     except Book.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+        return Response({"Message": "Student not found in DB"}, status=status.HTTP_404_NOT_FOUND)
     serializer = StudentSerializer(student, data=request.data)
     if serializer.is_valid():
         serializer.save()
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -198,9 +227,18 @@ def get_all_books_with_filters(request):
     student_name = request.GET.get('student_name')
     book_name = request.GET.get('book_name')
     checkout_date = request.GET.get('checkout_date')
-    book = Book.objects.get(title=book_name)
-    student = Student.objects.get(name=student_name)
-    books_issued = BookIssue.objects.filter(book_id=book.pk, student_id=student.pk, issue_date=checkout_date)
+    try:
+        book = Book.objects.get(title=book_name)
+    except Book.DoesNotExist:
+        return Response({"Message": "Book not found in DB with these filters"}, status=status.HTTP_404_NOT_FOUND)
+    try:
+        student = Student.objects.get(name=student_name)
+    except Student.DoesNotExist:
+        return Response({"Message": "Student not found in DB with these filters"}, status=status.HTTP_404_NOT_FOUND)
+    try:
+        books_issued = BookIssue.objects.filter(book_id=book.pk, student_id=student.pk, issue_date=checkout_date)
+    except BookIssue.DoesNotExist:
+        return Response({"Message": "Book not issued already"}, status=status.HTTP_404_NOT_FOUND)
     books_checked_out = []
     for book_issued in books_issued:
         books_checked_out.append(book_issued.book.title)
